// stencil-overlay-test/src/utils/window-event-map.enum.ts
export enum WindowEventMap {
	abort = 'abort',
	afterprint = 'afterprint',
	beforeprint = 'beforeprint',
	beforeunload = 'beforeunload',
	blur = 'blur',
	canplay = 'canplay',
	canplaythrough = 'canplaythrough',
	change = 'change',
	click = 'click',
	compassneedscalibration = 'compassneedscalibration',
	contextmenu = 'contextmenu',
	dblclick = 'dblclick',
	devicelight = 'devicelight',
	devicemotion = 'devicemotion',
	deviceorientation = 'deviceorientation',
	deviceorientationabsolute = 'deviceorientationabsolute',
	drag = 'drag',
	dragend = 'dragend',
	dragenter = 'dragenter',
	dragleave = 'dragleave',
	dragover = 'dragover',
	dragstart = 'dragstart',
	drop = 'drop',
	durationchange = 'durationchange',
	emptied = 'emptied',
	ended = 'ended',
	error = 'error',
	focus = 'focus',
	hashchange = 'hashchange',
	input = 'input',
	invalid = 'invalid',
	keydown = 'keydown',
	keypress = 'keypress',
	keyup = 'keyup',
	load = 'load',
	loadeddata = 'loadeddata',
	loadedmetadata = 'loadedmetadata',
	loadstart = 'loadstart',
	message = 'message',
	mousedown = 'mousedown',
	mouseenter = 'mouseenter',
	mouseleave = 'mouseleave',
	mousemove = 'mousemove',
	mouseout = 'mouseout',
	mouseover = 'mouseover',
	mouseup = 'mouseup',
	mousewheel = 'mousewheel',
	MSGestureChange = 'MSGestureChange',
	MSGestureDoubleTap = 'MSGestureDoubleTap',
	MSGestureEnd = 'MSGestureEnd',
	MSGestureHold = 'MSGestureHold',
	MSGestureStart = 'MSGestureStart',
	MSGestureTap = 'MSGestureTap',
	MSInertiaStart = 'MSInertiaStart',
	MSPointerCancel = 'MSPointerCancel',
	MSPointerDown = 'MSPointerDown',
	MSPointerEnter = 'MSPointerEnter',
	MSPointerLeave = 'MSPointerLeave',
	MSPointerMove = 'MSPointerMove',
	MSPointerOut = 'MSPointerOut',
	MSPointerOver = 'MSPointerOver',
	MSPointerUp = 'MSPointerUp',
	offline = 'offline',
	online = 'online',
	orientationchange = 'orientationchange',
	pagehide = 'pagehide',
	pageshow = 'pageshow',
	pause = 'pause',
	play = 'play',
	playing = 'playing',
	popstate = 'popstate',
	progress = 'progress',
	ratechange = 'ratechange',
	readystatechange = 'readystatechange',
	reset = 'reset',
	resize = 'resize',
	scroll = 'scroll',
	seeked = 'seeked',
	seeking = 'seeking',
	select = 'select',
	stalled = 'stalled',
	storage = 'storage',
	submit = 'submit',
	suspend = 'suspend',
	timeupdate = 'timeupdate',
	unload = 'unload',
	volumechange = 'volumechange',
	vrdisplayactivate = 'vrdisplayactivate',
	vrdisplayblur = 'vrdisplayblur',
	vrdisplayconnect = 'vrdisplayconnect',
	vrdisplaydeactivate = 'vrdisplaydeactivate',
	vrdisplaydisconnect = 'vrdisplaydisconnect',
	vrdisplayfocus = 'vrdisplayfocus',
	vrdisplaypointerrestricted = 'vrdisplaypointerrestricted',
	vrdisplaypointerunrestricted = 'vrdisplaypointerunrestricted',
	vrdisplaypresentchange = 'vrdisplaypresentchange',
	waiting = 'waiting',
}